+++ products/jbrowse-cli/src/commands/add-connection.test.ts
@@ -5,36 +5,11 @@
 import fs from 'fs'
 import path from 'path'
 
-import { setup, readConf, runInTmpDir } from '../testUtil'
+import { readConf, runInTmpDir } from '../testUtil'
 import nock from 'nock'
 import { runCommand } from '@oclif/test'
 
 const { copyFile, rename } = fs.promises
-
-const defaultConfig = {
-  assemblies: [
-    {
-      name: 'testAssembly',
-      sequence: {
-        type: 'testSequenceTrack',
-        trackId: '',
-        adapter: {
-          type: 'testSeqAdapter',
-          twoBitLocation: {
-            uri: 'test.2bit',
-            locationType: 'UriLocation',
-          },
-        },
-      },
-    },
-  ],
-  configuration: {},
-  connections: [],
-  defaultSession: {
-    name: 'New Session',
-  },
-  tracks: [],
-}
 
 const testConfig = path.join(
   __dirname,
@@ -53,17 +28,9 @@ async function copyConf(ctx: { dir: string }) {
     path.join(ctx.dir, 'config.json'),
   )
 }
-// const originalDateNow = Date.now
 
-// const setupWithDateMock = setup
-//   .do(() => {
-//     Date.now = jest.fn(() => 1)
-//   })
-//   .finally(() => {
-//     Date.now = originalDateNow
-//   })
-
-// Cleaning up exitCode in Node.js 20, xref https://github.com/jestjs/jest/issues/14501
+// Cleaning up exitCode in Node.js 20, xref
+// https://github.com/jestjs/jest/issues/14501
 afterAll(() => (process.exitCode = 0))
 beforeAll(() => (Date.now = jest.fn(() => 1)))
 
@@ -71,22 +38,24 @@ test('fails if no config file', async () => {
   nock('https://example.com').head('/hub.txt').reply(200)
 
   const { error } = await runCommand([
-    'add-connection https://example.com/hub.txt',
+    'add-connection',
+    'https://example.com/hub.txt',
   ])
-  expect(error).toMatchSnapshot()
+  expect(error?.message).toMatchSnapshot()
 })
 
 test('fails if data directory is not an url', async () => {
   const { error } = await runCommand(['add-connection .'])
-  expect(error).toMatchSnapshot()
+  expect(error?.message).toMatchSnapshot()
 })
 
 test('fails when fetching from url fails', async () => {
   nock('https://mysite.com').head('/notafile.txt').reply(500)
   const { error } = await runCommand([
-    'add-connection https://mysite.com/notafile.txt',
+    'add-connection',
+    'https://mysite.com/notafile.txt',
   ])
-  expect(error).toMatchSnapshot()
+  expect(error?.message).toMatchSnapshot()
 })
 
 test('adds an UCSCTrackHubConnection connection from a url', async () => {
@@ -94,7 +63,7 @@ test('adds an UCSCTrackHubConnection connection from a url', async () => {
     nock('https://mysite.com').head('/data/hub.txt').reply(200)
     await copyConf(ctx)
     await runCommand(['add-connection', 'https://mysite.com/data/hub.txt'])
-    expect(readConf(ctx)).toMatchSnapshot()
+    expect(readConf(ctx).connections).toMatchSnapshot()
   })
 })
 
@@ -103,7 +72,7 @@ test('adds JBrowse1 connection from a url', async () => {
     nock('https://mysite.com').head('/jbrowse/data').reply(200)
     await copyConf(ctx)
     await runCommand(['add-connection', 'https://mysite.com/jbrowse/data'])
-    expect(readConf(ctx)).toMatchSnapshot()
+    expect(readConf(ctx).connections).toMatchSnapshot()
   })
 })
 test('adds a custom connection with user set fields', async () => {
@@ -124,7 +93,7 @@ test('adds a custom connection with user set fields', async () => {
       '--config',
       '{"url":{"uri":"https://mysite.com/custom"},"locationType":"UriLocation"}',
     ])
-    expect(readConf(ctx)).toMatchSnapshot()
+    expect(readConf(ctx).connections).toMatchSnapshot()
   })
 })
 test('fails to add a duplicate connection', async () => {
@@ -148,7 +117,7 @@ test('fails to add a duplicate connection', async () => {
       '--config',
       '{"url":{"uri":"https://mysite.com/custom"},"locationType":"UriLocation"}',
     ])
-    expect(error).toMatchSnapshot()
+    expect(error?.message).toMatchSnapshot()
   })
 })
 test('overwrites an existing custom connection and does not check URL', async () => {
@@ -174,6 +143,6 @@ test('overwrites an existing custom connection and does not check URL', async ()
       '{"url":{"uri":"https://mysite.com/custom"},"locationType":"UriLocation"}',
       '--force',
     ])
-    expect(readConf(ctx)).toMatchSnapshot()
+    expect(readConf(ctx).connections).toMatchSnapshot()
   })
 })
