+++ products/jbrowse-cli/src/commands/admin-server.test.ts
@@ -9,7 +9,6 @@ import { runCommand } from '@oclif/test'
 // locals
 import fetch from '../fetchWithProxy'
 import { dataDir, readConf, runInTmpDir } from '../testUtil'
-import { getAdminKey, killExpress } from './admin-server-test-utils'
 
 const { copyFile, rename, chmod } = fs.promises
 
@@ -21,6 +20,36 @@ const testIndex = dataDir('simpleIndex.html')
 // Cleaning up exitCode in Node.js 20, xref
 // https://github.com/jestjs/jest/issues/14501
 afterAll(() => (process.exitCode = 0))
+
+function getPort(output: string) {
+  const portMatch = output.match(/localhost:([0-9]{4})/)
+  const port = portMatch?.[1]
+  if (!port) {
+    throw new Error(`Port not found in "${JSON.stringify(output)}"`)
+  }
+  return port
+}
+
+function getAdminKey(output: string) {
+  const keyMatch = output.match(/adminKey=([a-zA-Z0-9]{10,12}) /)
+  const key = keyMatch?.[1]
+  if (!key) {
+    throw new Error(`Admin key not found in "${output}"`)
+  }
+  return key
+}
+
+async function killExpress({ stdout }: { stdout: string }) {
+  // if (!stdout || typeof stdout !== 'string') {
+  //   // This test didn't start a server
+  //   return
+  // }
+  return fetch(`http://localhost:${getPort(stdout)}/shutdown`, {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({ adminKey: getAdminKey(stdout) }),
+  })
+}
 
 test('creates a default config', async () => {
   await runInTmpDir(async ctx => {
@@ -60,14 +89,14 @@ test('uses port 9090 if not specified', async () => {
 test('throws an error with a negative port', async () => {
   await runInTmpDir(async () => {
     const { error } = await runCommand(['admin-server', '--port', '-10'])
-    expect(error).toMatchSnapshot()
+    expect(error?.message).toMatchSnapshot()
   })
 })
 
 test('throws an error with a port greater than 65535', async () => {
   await runInTmpDir(async () => {
     const { error } = await runCommand(['admin-server', '--port', '66666'])
-    expect(error).toMatchSnapshot()
+    expect(error?.message).toMatchSnapshot()
   })
 })
